Init customTokenSource.refreshTime
